# packages/api/tsoa-build/swagger-spec.json
{
	"components": {
...
		"schemas": {
			"TutorId": {
				"type": "string",
				"description": "A tutor user id"
			},
			"StudentId": {
				"type": "string",
				"description": "A student user id"
			},
			"UserId": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/TutorId"
					},
					{
						"$ref": "#/components/schemas/StudentId"
					}
				],
				"description": "A general user id, ie a student or tutor id"
			},
			"UrlString": {
				"type": "string",
				"description": "An internet link"
			},
			"Country": {
				"type": "string",
				"enum": [
					"American Samoa",
					"Anguilla",
					"Antigua and Barbuda",
					"Australia",
					"Austria",
					"Bahamas",
					"Barbados",
					"Belgium",
					"Belize",
					"Bermuda",
					"Botswana",
					"British Indian Ocean Territory",
					"British Virgin Islands",
					"Burundi",
					"Cameroon",
					"Canada",
					"Caribbean",
					"Cayman Islands",
					"Christmas Island",
					"Cocos [Keeling] Islands",
					"Cook Islands",
					"Cyprus",
					"Denmark",
					"Dominica",
					"Eritrea",
					"Europe",
					"Falkland Islands",
					"Finland",
					"Fiji",
					"Gambia",
					"Germany",
					"Ghana",
					"Gibraltar",
					"Grenada",
					"Guam",
					"Guernsey",
					"Guyana",
					"Hong Kong",
					"India",
					"Ireland",
					"Isle of Man",
					"Israel",
					"Jamaica",
					"Jersey",
					"Kenya",
					"Kiribati",
					"Lesotho",
					"Liberia",
					"Macao SAR",
					"Madagascar",
					"Malawi",
					"Malaysia",
					"Malta",
					"Marshall Islands",
					"Mauritius",
					"Micronesia",
					"Montserrat",
					"Namibia",
					"Nauru",
					"Netherlands",
					"New Zealand",
					"Nigeria",
					"Niue",
					"Norfolk Island",
					"Northern Mariana Islands",
					"Pakistan",
					"Palau",
					"Papua New Guinea",
					"Pitcairn Islands",
					"Puerto Rico",
					"Republic of the Philippines",
					"Rwanda",
					"Saint Kitts and Nevis",
					"Saint Lucia",
					"Saint Vincent and the Grenadines",
					"Samoa",
					"Seychelles",
					"Sierra Leone",
					"Singapore",
					"Sint Maarten",
					"Slovenia",
					"Solomon Islands",
					"South Africa",
					"South Sudan",
					"St Helena, Ascension,  Tristan da Cunha",
					"Sudan",
					"Swaziland",
					"Sweden",
					"Switzerland",
					"Tanzania",
					"Tokelau",
					"Tonga",
					"Trinidad and Tobago",
					"Turks and Caicos Islands",
					"Tuvalu",
					"Uganda",
					"United Arab Emirates",
					"United Kingdom",
					"United States",
					"US Minor Outlying Islands",
					"US Virgin Islands",
					"Vanuatu",
					"World",
					"Zambia",
					"Zimbabwe",
					"Algeria",
					"Benin",
					"Burkina Faso",
					"Central African Republic",
					"Chad",
					"Comoros",
					"Congo",
					"Congo, DRC",
					"Côte d'Ivoire",
					"Djibouti",
					"Equatorial Guinea",
					"France",
					"French Guiana",
					"French Polynesia",
					"Gabon",
					"Guadeloupe",
					"Guinea",
					"Haiti",
					"Luxembourg",
					"Mali",
					"Martinique",
					"Mauritania",
					"Mayotte",
					"Morocco",
					"New Caledonia",
					"Niger",
					"Principality of Monaco",
					"Reunion",
					"Saint Barthélemy",
					"Saint Martin",
					"Saint Pierre and Miquelon",
					"Senegal",
					"Syria",
					"Togo",
					"Tunisia",
					"Wallis and Futuna",
					"Brunei Darussalam",
					"Angola",
					"Brazil",
					"Cabo Verde",
					"Guinea-Bissau",
					"Mozambique",
					"Portugal",
					"São Tomé and Príncipe",
					"Timor-Leste"
				],
				"description": "A country name e.g. \"Malaysia\""
			},
			"LocaleCode": {
				"type": "string",
				"enum": [
					"en",
					"fr",
					"pt",
					"ms"
				],
				"description": "A locale code to specify language e.g. en for English"
			},
			"PublicStudentData": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/UserId",
						"description": "Unique id for a user"
					},
					"imageUrl": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UrlString"
							}
						],
						"nullable": true,
						"description": "Link to profile image, if defined"
					},
					"introduction": {
						"type": "string",
						"description": "Optional text for users to introduce themselves,\r\ndescribe their experience, what subjects they are interested in etc"
					},
					"prefferedCountries": {
						"items": {
							"$ref": "#/components/schemas/Country"
						},
						"type": "array",
						"description": "The countries the user prefers for content"
					},
					"prefferedLocales": {
						"items": {
							"$ref": "#/components/schemas/LocaleCode"
						},
						"type": "array",
						"description": "The locales a user prefers for content"
					},
					"userName": {
						"type": "string",
						"description": "Username which will be displayed for a user"
					}
				},
				"required": [
					"id",
					"prefferedCountries",
					"prefferedLocales",
					"userName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SubjectId": {
				"type": "string",
				"description": "A general subject id that doesnt specify locale or country"
			},
			"SubjectName": {
				"type": "string",
				"description": "Name of a subject in a specific locale"
			},
			"LocaleSubjectData": {
				"description": "Schema of the general public data (not relating to a user) that represents a specific locale dependent subject,\r\nas it is stored in database. e.g. \"Physics\" and \"Physiques\" are two different LocaleSubjects\r\n// ! this is represented as separate documents as there might be a lot of data associated with locale subjects, so this prevents every locale subject being loaded when querying generic subjects",
				"properties": {
					"id": {
						"$ref": "#/components/schemas/SubjectId",
						"description": "Id of generic subject this locale subject relates to, there should only be 1"
					},
					"country": {
						"$ref": "#/components/schemas/Country",
						"description": "Specific country for this subject, for any subjects that might be country dependent, blank means it applies to any"
					},
					"description": {
						"type": "string",
						"description": "Description of the subject"
					},
					"locale": {
						"$ref": "#/components/schemas/LocaleCode",
						"description": "Represents a language that a subject is in"
					},
					"name": {
						"$ref": "#/components/schemas/SubjectName"
					},
					"relatedStudents": {
						"items": {
							"$ref": "#/components/schemas/StudentId"
						},
						"type": "array",
						"description": "Ids of students needing help with this subject `// todo needs to be syncronised`"
					},
					"relatedTutors": {
						"items": {
							"$ref": "#/components/schemas/TutorId"
						},
						"type": "array",
						"description": "Ids of tutors available to help with this subject `// todo needs to be syncronised`"
					}
				},
				"required": [
					"id",
					"country",
					"description",
					"locale",
					"name",
					"relatedStudents",
					"relatedTutors"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LinkStudentAndSubjectResponseBody": {
				"properties": {
					"student": {
						"$ref": "#/components/schemas/PublicStudentData"
					},
					"subject": {
						"$ref": "#/components/schemas/LocaleSubjectData"
					}
				},
				"required": [
					"student",
					"subject"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ConfidenceLevelEnum": {
				"description": "Confidence level options mapped to integers",
				"enum": [
					0,
					1,
					2,
					3
				],
				"type": "number"
			},
			"UserSubjectData": {
				"description": "Data about a subject specific to a user, as it is stored in database\r\n// ? should this be public?",
				"properties": {
					"id": {
						"$ref": "#/components/schemas/SubjectId",
						"description": "Unique id to the related subject"
					},
					"confidenceLevel": {
						"$ref": "#/components/schemas/ConfidenceLevelEnum",
						"description": "A rating of how confident a user feels in the subject, the numbers"
					},
					"country": {
						"$ref": "#/components/schemas/Country"
					},
					"detail": {
						"type": "string",
						"description": "Detail of the users experience with the subject"
					},
					"locale": {
						"$ref": "#/components/schemas/LocaleCode"
					}
				},
				"required": [
					"id",
					"confidenceLevel",
					"country",
					"detail",
					"locale"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_LinkStudentAndSubjectRequestBody_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/UserSubjectData"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"PublicTutorData": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/UserId",
						"description": "Unique id for a user"
					},
					"imageUrl": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UrlString"
							}
						],
						"nullable": true,
						"description": "Link to profile image, if defined"
					},
					"introduction": {
						"type": "string",
						"description": "Optional text for users to introduce themselves,\r\ndescribe their experience, what subjects they are interested in etc"
					},
					"prefferedCountries": {
						"items": {
							"$ref": "#/components/schemas/Country"
						},
						"type": "array",
						"description": "The countries the user prefers for content"
					},
					"prefferedLocales": {
						"items": {
							"$ref": "#/components/schemas/LocaleCode"
						},
						"type": "array",
						"description": "The locales a user prefers for content"
					},
					"userName": {
						"type": "string",
						"description": "Username which will be displayed for a user"
					},
					"available": {
						"type": "boolean",
						"description": "Represents whether a tutor is available to take on a student"
					}
				},
				"required": [
					"id",
					"prefferedCountries",
					"prefferedLocales",
					"userName",
					"available"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LinkStudentAndTutorResponseBody": {
				"properties": {
					"student": {
						"$ref": "#/components/schemas/PublicStudentData"
					},
					"tutor": {
						"$ref": "#/components/schemas/PublicTutorData"
					}
				},
				"required": [
					"student",
					"tutor"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_LinkStudentAndTutorRequestBody_": {
				"properties": {
					"studentId": {
						"$ref": "#/components/schemas/StudentId"
					},
					"tutorId": {
						"$ref": "#/components/schemas/TutorId"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"LinkSubjectAndSubjectCategoryResponseBody": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"CategoryId": {
				"type": "string",
				"description": "A general category name which doesnt specify locale"
			},
			"Partial_LinkSubjectAndSubjectCategoryRequestBody_": {
				"properties": {
					"categoryId": {
						"$ref": "#/components/schemas/CategoryId"
					},
					"locale": {
						"$ref": "#/components/schemas/LocaleCode",
						"description": "Locale to be returned after linking"
					},
					"subjectId": {
						"$ref": "#/components/schemas/SubjectId"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"LinkSubjectsResponseBody": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"Partial_LinkSubjectsRequestBody_": {
				"properties": {
					"subject1Id": {
						"$ref": "#/components/schemas/SubjectId"
					},
					"subject2Id": {
						"$ref": "#/components/schemas/SubjectId"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"LinkedStudentData": {
				"description": "Data about a student from the perspective of a tutor, as it is stored in database",
				"properties": {
					"id": {
						"$ref": "#/components/schemas/StudentId",
						"description": "Id of the student related to the tutor"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EmailString": {
				"type": "string",
				"description": "A valid email address"
			},
			"PrivateTutorData": {
				"description": "Schema of data that represents a tutor, only available to the tutor",
				"properties": {
					"id": {
						"$ref": "#/components/schemas/UserId",
						"description": "Unique id for a user"
					},
					"imageUrl": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UrlString"
							}
						],
						"nullable": true,
						"description": "Link to profile image, if defined"
					},
					"introduction": {
						"type": "string",
						"description": "Optional text for users to introduce themselves,\r\ndescribe their experience, what subjects they are interested in etc"
					},
					"prefferedCountries": {
						"items": {
							"$ref": "#/components/schemas/Country"
						},
						"type": "array",
						"description": "The countries the user prefers for content"
					},
					"prefferedLocales": {
						"items": {
							"$ref": "#/components/schemas/LocaleCode"
						},
						"type": "array",
						"description": "The locales a user prefers for content"
					},
					"userName": {
						"type": "string",
						"description": "Username which will be displayed for a user"
					},
					"email": {
						"$ref": "#/components/schemas/EmailString",
						"description": "Email to allow users to start communicating"
					},
					"relatedSubjects": {
						"items": {
							"$ref": "#/components/schemas/UserSubjectData"
						},
						"type": "array",
						"description": "Links to subjects a user is interested in learning/teaching (depends on type of user)\r\nand user specific information about the user relating to a specific subject"
					},
					"available": {
						"type": "boolean",
						"description": "Represents whether a tutor is available to take on a student"
					},
					"relatedStudents": {
						"items": {
							"$ref": "#/components/schemas/LinkedStudentData"
						},
						"type": "array",
						"description": "Data about the students a teacher is involved with"
					}
				},
				"required": [
					"id",
					"prefferedCountries",
					"prefferedLocales",
					"userName",
					"email",
					"relatedSubjects",
					"available",
					"relatedStudents"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LinkTutorAndSubjectResponseBody": {
				"properties": {
					"subject": {
						"$ref": "#/components/schemas/LocaleSubjectData"
					},
					"tutor": {
						"$ref": "#/components/schemas/PrivateTutorData"
					}
				},
				"required": [
					"subject",
					"tutor"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_LinkTutorAndSubjectRequestBody_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/UserSubjectData"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UnlinkStudentAndSubjectResponseBody": {
				"properties": {
					"student": {
						"$ref": "#/components/schemas/PublicStudentData"
					},
					"subject": {
						"$ref": "#/components/schemas/LocaleSubjectData"
					}
				},
				"required": [
					"student",
					"subject"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_UnlinkStudentAndSubjectRequestBody_": {
				"properties": {
					"country": {
						"$ref": "#/components/schemas/Country"
					},
					"id": {
						"$ref": "#/components/schemas/SubjectId"
					},
					"locale": {
						"$ref": "#/components/schemas/LocaleCode"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UnlinkStudentAndTutorResponseBody": {
				"properties": {
					"student": {
						"$ref": "#/components/schemas/PublicStudentData"
					},
					"tutor": {
						"$ref": "#/components/schemas/PublicTutorData"
					}
				},
				"required": [
					"student",
					"tutor"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_UnlinkStudentAndTutorRequestBody_": {
				"properties": {
					"studentId": {
						"$ref": "#/components/schemas/StudentId"
					},
					"tutorId": {
						"$ref": "#/components/schemas/TutorId"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UnlinkSubjectAndSubjectCategoryResponseBody": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"Partial_UnlinkSubjectAndSubjectCategoryRequestBody_": {
				"properties": {
					"categoryId": {
						"$ref": "#/components/schemas/CategoryId"
					},
					"subjectId": {
						"$ref": "#/components/schemas/SubjectId"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UnlinkSubjectsResponseBody": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"Partial_UnlinkSubjectsRequestBody_": {
				"properties": {
					"subject1Id": {
						"$ref": "#/components/schemas/SubjectId"
					},
					"subject2Id": {
						"$ref": "#/components/schemas/SubjectId"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UnlinkTutorAndSubjectResponseBody": {
				"properties": {
					"subject": {
						"$ref": "#/components/schemas/LocaleSubjectData"
					},
					"tutor": {
						"$ref": "#/components/schemas/PublicTutorData"
					}
				},
				"required": [
					"subject",
					"tutor"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_UnlinkTutorAndSubjectRequestBody_": {
				"properties": {
					"country": {
						"$ref": "#/components/schemas/Country"
					},
					"id": {
						"$ref": "#/components/schemas/SubjectId"
					},
					"locale": {
						"$ref": "#/components/schemas/LocaleCode"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@adopt-a-student/api",
		"version": "0.0.0",
		"description": "API for adopt-a-student",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Elias Mangoro"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/linkStudentAndSubject": {
			"post": {
				"operationId": "LinkStudentAndSubject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LinkStudentAndSubjectResponseBody"
								}
							}
						}
					}
				},
				"description": "Retreives data about a tutor user. If the tutor user owns the data then they get all the data, otherwise it is restricted to 'public' data.",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_LinkStudentAndSubjectRequestBody_"
							}
						}
					}
				}
			}
		},
		"/linkStudentAndTutor": {
			"post": {
				"operationId": "LinkStudentAndTutor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LinkStudentAndTutorResponseBody"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_LinkStudentAndTutorRequestBody_"
							}
						}
					}
				}
			}
		},
		"/linkSubjectAndSubjectCategory": {
			"post": {
				"operationId": "LinkSubjectAndSubjectCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LinkSubjectAndSubjectCategoryResponseBody"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_LinkSubjectAndSubjectCategoryRequestBody_"
							}
						}
					}
				}
			}
		},
		"/linkSubjects": {
			"post": {
				"operationId": "LinkSubjects",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LinkSubjectsResponseBody"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_LinkSubjectsRequestBody_"
							}
						}
					}
				}
			}
		},
		"/linkTutorAndSubject": {
			"post": {
				"operationId": "LinkTutorAndSubject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LinkTutorAndSubjectResponseBody"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_LinkTutorAndSubjectRequestBody_"
							}
						}
					}
				}
			}
		},
		"/unlinkStudentAndSubject": {
			"post": {
				"operationId": "UnlinkStudentAndSubject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnlinkStudentAndSubjectResponseBody"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_UnlinkStudentAndSubjectRequestBody_"
							}
						}
					}
				}
			}
		},
		"/unlinkStudentAndTutor": {
			"post": {
				"operationId": "UnlinkStudentAndTutor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnlinkStudentAndTutorResponseBody"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_UnlinkStudentAndTutorRequestBody_"
							}
						}
					}
				}
			}
		},
		"/unlinkSubjectAndSubjectCategory": {
			"post": {
				"operationId": "UnlinkSubjectAndSubjectCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnlinkSubjectAndSubjectCategoryResponseBody"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_UnlinkSubjectAndSubjectCategoryRequestBody_"
							}
						}
					}
				}
			}
		},
		"/unlinkSubjects": {
			"post": {
				"operationId": "UnlinkSubjects",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnlinkSubjectsResponseBody"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_UnlinkSubjectsRequestBody_"
							}
						}
					}
				}
			}
		},
		"/unlinkTutorAndSubject": {
			"post": {
				"operationId": "UnlinkTutorAndSubject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnlinkTutorAndSubjectResponseBody"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_UnlinkTutorAndSubjectRequestBody_"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}